+++ src/components/blog-details/BlogDetailsShimmerPage.tsx
